feat: create EventController to get user's input, refactor domain, service and repository of Event feature

+++ src/main/java/org/agenda/event/dto/UpdateEventRequest.java
@@ -6,6 +6,6 @@ import org.jetbrains.annotations.Nullable;
 
 import java.time.LocalDateTime;
 
-public record UpdateEventRequest(int id, @Nullable String title, @Nullable String description, @Nullable LocalDateTime date, @Nullable String type, @Nullable String eventSchedule) {
+public record UpdateEventRequest(long id, @Nullable String title, @Nullable String description, @Nullable LocalDateTime date, @Nullable String type, @Nullable String eventSchedule) {
 
 }
